requirements.txt
fastapi
uvicorn
httpx
pydantic
gunicorn
hazelcast-python-client
aiokafka
python-consul